refactor(SDK-1122): split Contractor Documents into onboarding/ and shared/
Aligns Contractor/Documents with the Employee domain's onboarding/management/shared
convention. DocumentSigner and the DocumentsList component are onboarding-only, so
they move under onboarding/; SignatureForm and the useContractorDocumentsList hook
are consumed by both onboarding and management, so they move under shared/ (matching
Employee.Documents.shared.SignatureForm / shared.useDocumentsList). Pure move plus
import-path updates; no behavior change.
