+++ docs/guides/endpoint-inventory.json
@@ -2200,11 +2200,17 @@
       "endpoints": [
         {
           "method": "GET",
+          "path": "/v1/contractors/:contractorUuid/documents",
+          "docsUrl": "https://docs.gusto.com/embedded-payroll/v2026-06-15/reference/get-v1-contractor-documents"
+        },
+        {
+          "method": "GET",
           "path": "/v1/documents/:documentUuid/pdf",
           "docsUrl": "https://docs.gusto.com/embedded-payroll/v2026-06-15/reference/get-v1-contractor-document-pdf"
         }
       ],
       "variables": [
+        "contractorUuid",
         "documentUuid"
       ]
     },
