--- src/components/Contractor/Documents/DocumentsList/index.ts
@@ -1,7 +1,0 @@
-export { DocumentsList, type DocumentsListProps } from './DocumentsList'
-export {
-  useContractorDocumentsList,
-  type UseContractorDocumentsListParams,
-  type UseContractorDocumentsListReady,
-  type UseContractorDocumentsListResult,
-} from './useContractorDocumentsList'
