+++ src/components/Contractor/Documents/management/DocumentsCard/DocumentsCard.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useContractorDocumentsGetPdf } from '@gusto/embedded-api/react-query/contractorDocumentsGetPdf'
 import type { Document } from '@gusto/embedded-api/models/components/document'
-import { useContractorDocumentsList } from '../../DocumentsList/useContractorDocumentsList'
+import { useContractorDocumentsList } from '../../shared/useContractorDocumentsList'
 import { DataView, EmptyData, useDataView, Loading } from '@/components/Common'
 import { BaseBoundaries, BaseLayout } from '@/components/Base/Base'
 import { useComponentContext } from '@/contexts/ComponentAdapter/useComponentContext'
