+++ src/components/Contractor/Documents/onboarding/DocumentsList/index.ts
@@ -1,0 +1,1 @@
+export { DocumentsList, type DocumentsListProps } from './DocumentsList'
