+++ src/components/Contractor/SelfOnboardingFlow/SelfOnboardingComponents.tsx
@@ -6,7 +6,7 @@ import { Landing as LandingComponent } from '@/components/Contractor/Landing'
 import { ContractorProfile as ProfileComponent } from '@/components/Contractor/Profile'
 import { Address as AddressComponent } from '@/components/Contractor/Address'
 import { PaymentMethod as PaymentMethodComponent } from '@/components/Contractor/PaymentMethod/PaymentMethod'
-import { DocumentSigner as DocumentSignerComponent } from '@/components/Contractor/Documents/DocumentSigner'
+import { DocumentSigner as DocumentSignerComponent } from '@/components/Contractor/Documents/onboarding/DocumentSigner'
 import { OnboardingSummary as OnboardingSummaryComponent } from '@/components/Contractor/OnboardingSummary'
 
 /**
