+++ src/components/Contractor/exports/contractorOnboarding.ts
@@ -27,6 +27,6 @@ export type { PaymentMethodProps } from '../PaymentMethod/types'
 export { NewHireReport } from '../NewHireReport/NewHireReport'
 export type { NewHireReportProps } from '../NewHireReport/types'
 export { ContractorSubmit, type ContractorSubmitProps } from '../Submit/Submit'
-export { DocumentSigner, type DocumentSignerProps } from '../Documents/DocumentSigner'
-export { DocumentsList, type DocumentsListProps } from '../Documents/DocumentsList'
-export { SignatureForm, type SignatureFormProps } from '../Documents/SignatureForm'
+export { DocumentSigner, type DocumentSignerProps } from '../Documents/onboarding/DocumentSigner'
+export { DocumentsList, type DocumentsListProps } from '../Documents/onboarding/DocumentsList'
+export { SignatureForm, type SignatureFormProps } from '../Documents/shared/SignatureForm'
