+++ src/index.ts
@@ -630,17 +630,17 @@ export type {
 } from '@/components/Company/DocumentSigner/shared/useSignCompanyForm'
 
 // Domain data hooks - Contractor
-export { useContractorDocumentsList } from '@/components/Contractor/Documents/DocumentsList/useContractorDocumentsList'
+export { useContractorDocumentsList } from '@/components/Contractor/Documents/shared/useContractorDocumentsList'
 export type {
   UseContractorDocumentsListParams,
   UseContractorDocumentsListResult,
   UseContractorDocumentsListReady,
-} from '@/components/Contractor/Documents/DocumentsList/useContractorDocumentsList'
+} from '@/components/Contractor/Documents/shared/useContractorDocumentsList'
 
 export {
   useContractorSignatureForm,
   ContractorSignatureFormErrorCodes,
-} from '@/components/Contractor/Documents/SignatureForm/useContractorSignatureForm'
+} from '@/components/Contractor/Documents/shared/SignatureForm/useContractorSignatureForm'
 export type {
   UseContractorSignatureFormProps,
   UseContractorSignatureFormResult,
@@ -673,4 +673,4 @@ export type {
   ContractorSignatureAgreeValidation,
   ContractorSignatureSsnValidation,
   ContractorSignatureEinValidation,
-} from '@/components/Contractor/Documents/SignatureForm/useContractorSignatureForm'
+} from '@/components/Contractor/Documents/shared/SignatureForm/useContractorSignatureForm'
